requirements.txt
Django>=4.2
Pillow~=11.2.1


pip~=25.0.1
asgiref~=3.8.1
sqlparse~=0.5.3
djangorestframework~=3.16.0